+++ lib/langtool_web/controllers/dashboard_controller.ex
@@ -1,9 +1,12 @@
 defmodule LangtoolWeb.DashboardController do
   use LangtoolWeb, :controller
+  alias LangtoolWeb.{DashboardPolicy}
 
   plug :check_auth when action in [:index]
 
   def index(conn, _) do
-    render conn, "index.html"
+    conn
+    |> authorize(:dashboard, :index?, nil)
+    |> render("index.html")
   end
 end