+++ lib/langtool_web/policies/dashboard_policy.ex
@@ -1,0 +1,5 @@
+defmodule LangtoolWeb.DashboardPolicy do
+  def index?(_, _) do
+    true
+  end
+end